Renamed Event to Delegate.
