+++ src/core/delegate.js
