+++ src/core/scene.js
@@ -4,7 +4,7 @@
 
 define( function ( require ) {
 
-    var Event = require( './event' );
+    var Delegate = require( './delegate' );
 
     /* Scene
      *
@@ -179,8 +179,8 @@ define( function ( require ) {
                 return null;
             };
 
-            // Events
-            var _entityAdded = new Event();
+            // Delegates
+            var _entityAdded = new Delegate();
             Object.defineProperty( this, 'entityAdded', {
                 get: function() {
                     return _entityAdded;
@@ -192,7 +192,7 @@ define( function ( require ) {
                 }
             };
 
-            var _entityRemoved = new Event();
+            var _entityRemoved = new Delegate();
             Object.defineProperty( this, 'entityRemoved', {
                 get: function() {
                     return _entityRemoved;
