+++ test/index.html
@@ -13,7 +13,7 @@
     <script src="core/engine.js"></script>
     <script src="core/lang.js"></script>
     <script src="core/timer.js"></script>
-    <script src="core/event.js"></script>
+    <script src="core/delegate.js"></script>
     <script src="core/entity.js"></script>
     <script src="core/component.js"></script>
     <script src="core/scene.js"></script>
